Added siosocks/io/twisted.py and test_twisted.py.
Right now, twisted.py is the same as asyncio.py,
just with types added for clarity, and with the server removed.
Twisted Deferreds interoperate w/ awaitable coroutines well,
so I'll base the twisted backend off of the async one.

test_twisted.py is the same as test_asyncio.py,
except it uses the async server but the twisted client,
since we don't need to write a twisted server (yet at least).
I also refactored the tests a bit and added types.